+++ entity/src/epilot/models/operations/getrelations.py
@@ -3,7 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from typing import Any, Optional
+from ..shared import relationitem as shared_relationitem
+from typing import Any, Optional, Union
 
 
 
@@ -28,7 +29,7 @@ class GetRelationsResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    get_relations_resp: Optional[list[Any]] = dataclasses.field(default=None)
+    get_relations_resp: Optional[list[Union[shared_relationitem.RelationItem, dict[str, Any]]]] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""