+++ entity/src/epilot/relations.py
@@ -3,7 +3,7 @@
 from .sdkconfiguration import SDKConfiguration
 from epilot import utils
 from epilot.models import errors, operations, shared
-from typing import Any, Optional
+from typing import Any, Optional, Union
 
 class Relations:
     r"""Entity Relationships"""
@@ -93,7 +93,7 @@ class Relations:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[list[Any]])
+                out = utils.unmarshal_json(http_res.text, Optional[list[Union[shared.RelationItem, dict[str, Any]]]])
                 res.get_relations_resp = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)